Unify DirectRouter and NoStdRouter into single Router type
Replace two separate router profiles with one `Router<I, R, N, S>` that
works on both std and no_std. Based on NoStdRouter design (heapless::Vec,
injectable RNG, token rotation on refresh).

Std-specific features behind #[cfg(feature = "std")]: closer WaitQueue
for worker lifecycle, get_nets(), Default impl with StdRng.

Backwards compatibility via type aliases:
- DirectRouter<I> = Router<I, StdRng, 64, 64>
- NoStdRouter<I, R, N, S> = Router<I, R, N, S>

Breaking: register_interface returns Result<u8> instead of Option<u64>,
interface ident is now u8 (was u64 for DirectRouter).

+++ crates/ergot/src/interface_manager/transports/tokio_serial.rs
@@ -87,7 +87,7 @@ pub async fn register_router<N, I>(
     outgoing_buffer_size: usize,
     liveness: Option<LivenessConfig>,
     state_notify: Option<Arc<WaitQueue>>,
-) -> Result<u64, RouterRegistrationError>
+) -> Result<u8, RouterRegistrationError>
 where
     I: Interface<Sink = Sink<StdQueue>>,
     N: NetStackHandle<Profile = DirectRouter<I>> + Send + 'static,
